Recording through multiple devices through network::working
